phase-9-container-native/api/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.9
redis==5.0.8
pyyaml==6.0.2
